Address pr feedback and rebase to latest changes
