+++ src/System.Runtime/tests/System.Runtime.Tests.csproj
@@ -131,6 +131,7 @@
     <Compile Include="System\Collections\Generic\KeyValuePairTests.netcoreapp.cs" />
     <Compile Include="System\LazyTests.netcoreapp.cs" />
     <Compile Include="System\StringTests.netcoreapp.cs" />
+    <Compile Include="System\TimeSpanTests.netcoreapp.cs" />
     <Compile Include="System\Runtime\CompilerServices\RuntimeHelpersTests.netcoreapp.cs" />
     <Compile Include="System\Runtime\CompilerServices\ConditionalWeakTableTests.netcoreapp.cs" />
     <Compile Include="System\UIntPtrTests.netcoreapp.cs" />
